% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_summary.R
\name{.add_pert_kind_and_mode}
\alias{.add_pert_kind_and_mode}
\title{Add perturbation kind, mode and dose unit from \code{pert_type}}
\usage{
.add_pert_kind_and_mode(df)
}
\arguments{
\item{df}{Data frame with a \code{pert_type} column.}
}
\value{
\code{df} with added \code{pert_kind}, \code{mode} and
  \code{dose_unit} columns.
}
\description{
Add perturbation kind, mode and dose unit from \code{pert_type}
}
\keyword{internal}
