% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.as_numeric}
\alias{.as_numeric}
\title{Coerce to numeric without warnings}
\usage{
.as_numeric(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Numeric vector; \code{NA} where \code{x} is not numeric.
}
\description{
Coerce to numeric without warnings
}
\keyword{internal}
