% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_summary.R
\name{.build_tech_view}
\alias{.build_tech_view}
\title{Full technical view with preferred columns first}
\usage{
.build_tech_view(res2, compound_col, score_col, p_col, padj_col)
}
\arguments{
\item{res2}{Annotated data frame.}

\item{compound_col, score_col, p_col, padj_col}{Column names.}
}
\value{
Data frame.
}
\description{
Full technical view with preferred columns first
}
\keyword{internal}
