% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camsum.R
\name{.camsum_read_meta_file}
\alias{.camsum_read_meta_file}
\title{Read columns of a tab-separated CMap metadata file once per session}
\usage{
.camsum_read_meta_file(path, select)
}
\arguments{
\item{path}{File path.}

\item{select}{Columns to read.}
}
\value{
data.frame.
}
\description{
Read columns of a tab-separated CMap metadata file once per session
}
\keyword{internal}
