% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_gsea0}
\alias{.fast_gsea0}
\title{Vectorised GSEA weight-0 score}
\usage{
.fast_gsea0(refMatrix, queryUp, queryDown, permuteNum, pAdjMethod)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{queryUp, queryDown}{Character vectors of gene symbols.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}
}
\description{
Vectorised GSEA weight-0 score
}
\keyword{internal}
