% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_gseaW}
\alias{.fast_gseaW}
\title{Vectorised weighted GSEA score}
\usage{
.fast_gseaW(refMatrix, queryUp, queryDown, permuteNum, pAdjMethod, weight)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{queryUp, queryDown}{Character vectors of gene symbols.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}

\item{weight}{Exponent applied to the absolute reference values.}
}
\description{
Vectorised weighted GSEA score
}
\keyword{internal}
