% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_gseaW_es}
\alias{.fast_gseaW_es}
\title{Weighted GSEA enrichment of sorted positions}
\usage{
.fast_gseaW_es(SP, W, k, G)
}
\arguments{
\item{SP}{k x profiles matrix of sorted positions.}

\item{W}{k x profiles matrix of the weights, sorted like \code{SP}.}

\item{k}{Size of the gene set.}

\item{G}{Number of genes.}
}
\value{
Numeric vector, one value per profile.
}
\description{
Weighted GSEA enrichment of sorted positions
}
\keyword{internal}
