% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_ks}
\alias{.fast_ks}
\title{Vectorised KS score}
\usage{
.fast_ks(refMatrix, queryUp, queryDown, permuteNum, pAdjMethod)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{queryUp, queryDown}{Character vectors of gene symbols.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}
}
\description{
Vectorised KS score
}
\keyword{internal}
