% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_ks_es}
\alias{.fast_ks_es}
\title{KS enrichment of sorted positions}
\usage{
.fast_ks_es(SP, k, G)
}
\arguments{
\item{SP}{k x profiles matrix of sorted positions.}

\item{k}{Size of the gene set.}

\item{G}{Number of genes.}
}
\value{
Numeric vector, one value per profile.
}
\description{
KS enrichment of sorted positions
}
\keyword{internal}
