% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_xcos}
\alias{.fast_xcos}
\title{Vectorised XCos score}
\usage{
.fast_xcos(refMatrix, query, topN, permuteNum, pAdjMethod)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{query}{Named numeric vector.}

\item{topN}{Number of top/bottom genes kept per profile.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}
}
\description{
Vectorised XCos score
}
\keyword{internal}
