% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_xsum}
\alias{.fast_xsum}
\title{Vectorised XSum score}
\usage{
.fast_xsum(refMatrix, queryUp, queryDown, topN, permuteNum, pAdjMethod)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{queryUp, queryDown}{Character vectors of gene symbols.}

\item{topN}{Number of top/bottom genes kept per profile.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}
}
\description{
Vectorised XSum score
}
\keyword{internal}
