% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_fast.R
\name{.fast_zhang}
\alias{.fast_zhang}
\title{Vectorised Zhang score}
\usage{
.fast_zhang(refMatrix, queryUp, queryDown, permuteNum, pAdjMethod)
}
\arguments{
\item{refMatrix}{Numeric matrix, genes x profiles.}

\item{queryUp, queryDown}{Character vectors of gene symbols.}

\item{permuteNum}{Number of permutations.}

\item{pAdjMethod}{P-value adjustment method.}
}
\description{
Vectorised Zhang score
}
\keyword{internal}
