% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.is_blank}
\alias{.is_blank}
\title{Whether strings are missing or empty}
\usage{
.is_blank(x)
}
\arguments{
\item{x}{Vector coercible to character.}
}
\value{
Logical vector.
}
\description{
Whether strings are missing or empty
}
\keyword{internal}
