% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.matrix_to_extreme_list}
\alias{.matrix_to_extreme_list}
\title{Convert expression matrix to top-N / bottom-N named-value lists (for XCos/XSum)}
\usage{
.matrix_to_extreme_list(refMatrix, topN)
}
\description{
Convert expression matrix to top-N / bottom-N named-value lists (for XCos/XSum)
}
\keyword{internal}
