% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.matrix_to_value_ranked_list}
\alias{.matrix_to_value_ranked_list}
\title{Convert expression matrix to sorted named-value lists (for GSEA)}
\usage{
.matrix_to_value_ranked_list(refMatrix)
}
\arguments{
\item{refMatrix}{Numeric matrix with genes as rows and samples as columns.}
}
\value{
List of named numeric vectors ordered decreasingly by expression for each sample.
}
\description{
Convert expression matrix to sorted named-value lists (for GSEA)
}
\keyword{internal}
