% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_summary.R
\name{.merge_siginfo}
\alias{.merge_siginfo}
\title{Merge signature results with siginfo annotations}
\usage{
.merge_siginfo(res, sig_info_file, compound_col)
}
\arguments{
\item{res}{Results data frame.}

\item{sig_info_file}{Siginfo path or data frame.}

\item{compound_col}{Signature id column in \code{res}.}
}
\value{
Merged data frame with a guaranteed \code{pert_id} column.
}
\description{
Merge signature results with siginfo annotations
}
\keyword{internal}
