% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.message_rule}
\alias{.message_rule}
\title{Print a horizontal rule of \code{n} copies of \code{char} as a message}
\usage{
.message_rule(char = "=", n = 80L)
}
\arguments{
\item{char}{Single character.}

\item{n}{Width.}
}
\value{
NULL, invisibly.
}
\description{
Print a horizontal rule of \code{n} copies of \code{char} as a message
}
\keyword{internal}
