% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_siginfo_beta.R
\name{.print_reuse_filters}
\alias{.print_reuse_filters}
\title{Print the filters that reproduce the interactive selection}
\usage{
.print_reuse_filters(filtered_data, sig_info, key_columns)
}
\arguments{
\item{filtered_data}{Filtered data frame.}

\item{sig_info}{Unfiltered data frame.}

\item{key_columns}{Filter columns.}
}
\value{
NULL, invisibly.
}
\description{
Print the filters that reproduce the interactive selection
}
\keyword{internal}
