% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_siginfo_beta.R
\name{.read_siginfo}
\alias{.read_siginfo}
\title{Read a siginfo file}
\usage{
.read_siginfo(siginfo_file, verbose)
}
\arguments{
\item{siginfo_file}{Path to the siginfo table.}

\item{verbose}{Logical; print progress messages.}
}
\value{
Data frame.
}
\description{
Read a siginfo file
}
\keyword{internal}
