% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_siginfo_beta.R
\name{.remaining_columns}
\alias{.remaining_columns}
\title{Columns that follow position \code{col_idx} in \code{key_columns}}
\usage{
.remaining_columns(key_columns, col_idx)
}
\arguments{
\item{key_columns}{Character vector of filter columns.}

\item{col_idx}{Current position.}
}
\value{
Character vector (possibly empty).
}
\description{
Columns that follow position \code{col_idx} in \code{key_columns}
}
\keyword{internal}
