% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.validate_ref_matrix}
\alias{.validate_ref_matrix}
\title{Validate and coerce reference matrix}
\usage{
.validate_ref_matrix(refMatrix)
}
\description{
Validate and coerce reference matrix
}
\keyword{internal}
