% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_combinations.R
\name{generate_combinations_from_config}
\alias{generate_combinations_from_config}
\title{Generate combinations of time points, dosages, and cell lines from a configuration file}
\usage{
generate_combinations_from_config(config_file, verbose = TRUE)
}
\arguments{
\item{config_file}{Path to configuration file with time, dose, and cell options}

\item{verbose}{Logical; whether to print progress messages (default: TRUE)}
}
\value{
Data frame of combinations
}
\description{
Generate combinations of time points, dosages, and cell lines from a configuration file
}
\examples{
cfg <- create_cmap_config_template(dest_dir = tempdir(), overwrite = TRUE)
comb <- generate_combinations_from_config(cfg, verbose = FALSE)
head(comb)

\donttest{
# First create a configuration file
siginfo_file <- system.file("extdata", "example_siginfo.txt", package = "CONCERTDR")
extract_cmap_parameters(siginfo_file, write_config = TRUE)

# Then edit the config file to select desired parameters
# Leave sections empty to use all options for that parameter
# Now generate combinations from the edited config file
combinations <- generate_combinations_from_config("conf/cmap_options.conf")
}

}
