% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_utils.R
\name{generate_combinations_from_selections}
\alias{generate_combinations_from_selections}
\title{Generate combinations from selected parameters}
\usage{
generate_combinations_from_selections(selections, verbose = TRUE)
}
\arguments{
\item{selections}{The list returned by create_console_config()}

\item{verbose}{Logical; whether to print progress messages (default: TRUE)}
}
\value{
Data frame of combinations
}
\description{
Creates a data frame of combinations based on the selections from the interactive config functions
}
\examples{
sel <- list(
  selected = list(times = c("6 h"), doses = c("10 uM"), cells = c("A375", "MCF7")),
  all_options = list(times = c("6 h", "24 h"), doses = c("10 uM"), cells = c("A375", "MCF7"))
)
generate_combinations_from_selections(sel, verbose = FALSE)

\donttest{
siginfo_file <- system.file("extdata", "example_siginfo.txt", package = "CONCERTDR")
selections <- create_console_config(siginfo_file)

# Generate combinations
combinations <- generate_combinations_from_selections(selections)
}

}
