% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{get_rid}
\alias{get_rid}
\title{Get Unique Gene IDs and Gene Names}
\usage{
get_rid(geneinfo_df, landmark = TRUE)
}
\arguments{
\item{geneinfo_df}{Data frame containing gene information}

\item{landmark}{Restrict to landmark genes before resolving duplicate symbols.}
}
\value{
List with rid (gene IDs) and genenames (gene symbols)
}
\description{
Get Unique Gene IDs and Gene Names
}
\keyword{internal}
