% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_combinations.R
\name{read_cmap_config}
\alias{read_cmap_config}
\title{Read a CMap configuration file}
\usage{
read_cmap_config(config_file, return_selected = TRUE)
}
\arguments{
\item{config_file}{Path to configuration file}

\item{return_selected}{Whether to return selected values (default: TRUE)}
}
\value{
List with parsed configuration (all options and selected options)
}
\description{
Read a CMap configuration file
}
\examples{
cfg_file <- create_cmap_config_template(dest_dir = tempdir(), overwrite = TRUE)
cfg      <- read_cmap_config(cfg_file)
names(cfg)

}
