% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.apply_score}
\alias{.apply_score}
\title{Apply a scoring function across reference lists, optionally in parallel}
\usage{
.apply_score(refList, scoreFun, ...)
}
\description{
Apply a scoring function across reference lists, optionally in parallel
}
\keyword{internal}
