% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.matrix_to_name_ranked_list}
\alias{.matrix_to_name_ranked_list}
\title{Convert expression matrix to sorted gene-name lists (for KS / GSEA-w0)}
\usage{
.matrix_to_name_ranked_list(refMatrix)
}
\description{
Convert expression matrix to sorted gene-name lists (for KS / GSEA-w0)
}
\keyword{internal}
