% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.matrix_to_signed_rank_list}
\alias{.matrix_to_signed_rank_list}
\title{Convert expression matrix to signed-rank lists (for Zhang)}
\usage{
.matrix_to_signed_rank_list(refMatrix)
}
\description{
Convert expression matrix to signed-rank lists (for Zhang)
}
\keyword{internal}
