% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_methods.R
\name{.matrix_to_value_ranked_list}
\alias{.matrix_to_value_ranked_list}
\title{Convert expression matrix to sorted named-value lists (for GSEA-w1/w2)}
\usage{
.matrix_to_value_ranked_list(refMatrix)
}
\description{
Convert expression matrix to sorted named-value lists (for GSEA-w1/w2)
}
\keyword{internal}
