% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{fast_parse_gctx}
\alias{fast_parse_gctx}
\title{Fast parser for GCTX matrices with optional row/column subsetting}
\usage{
fast_parse_gctx(fname, rid = NULL, cid = NULL)
}
\arguments{
\item{fname}{Path to GCTX file}

\item{rid}{Optional character vector of row ids (gene ids)}

\item{cid}{Optional character vector of column ids (signature ids)}
}
\value{
Numeric matrix with selected rows/columns
}
\description{
Fast parser for GCTX matrices with optional row/column subsetting
}
\keyword{internal}
