% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_extract.R
\name{fast_read_meta}
\alias{fast_read_meta}
\title{Read first existing HDF5 dataset from candidate paths}
\usage{
fast_read_meta(fname, paths)
}
\arguments{
\item{fname}{HDF5/GCTX file path}

\item{paths}{Candidate dataset paths}
}
\value{
Dataset content
}
\description{
Read first existing HDF5 dataset from candidate paths
}
\keyword{internal}
